fix layout hooks
