+++ web/app/layout.tsx
@@ -1,10 +1,9 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import Link from "next/link";
-import { useEffect, useState } from "react";
 import "./globals.css";
 import SupabaseProvider from "@/components/SupabaseProvider";
-import { createClient } from "@/lib/supabaseClient";
+import QueueLink from "@/components/QueueLink";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -50,26 +49,3 @@ export default function RootLayout({
   );
 }
 
-function QueueLink() {
-  const supabase = createClient();
-  const [count, setCount] = useState(0);
-
-  useEffect(() => {
-    supabase
-      .from("block_change_queue")
-      .select("id", { count: "exact" })
-      .eq("status", "pending")
-      .then(({ count }) => setCount(count || 0));
-  }, []);
-
-  return (
-    <Link href="/blocks/queue" className="relative">
-      Queue
-      {count > 0 && (
-        <span className="absolute -right-3 -top-2 rounded-full bg-red-600 text-white text-xs px-1">
-          {count}
-        </span>
-      )}
-    </Link>
-  );
-}
